Work on the mem range logging.
The numbers in the serial log seem weird and probably are wrong. Upped
the digit converting to 64 bits to make sure to get all the int sizes.

+++ kernel_log.c
@@ -16,7 +16,7 @@ void kWriteLog(const char* str)
     }
 }
 
-void kWriteLog_format1d(const char* str, uint32_t number)
+void kWriteLog_format1d(const char* str, uint64_t number)
 {
     char outBuf[256];
