+++ kernel_log.h
@@ -3,12 +3,15 @@
 
 #include "serial.h"
 
+#include <stddef.h>
+#include <stdint.h>
+
 COMPRT keLogPort;
 
 void kSetupLog(COMPRT loggingPort);
 
 void kWriteLog(const char* str);
 
-void kWriteLog_format1d(const char* str, uint32_t number);
+void kWriteLog_format1d(const char* str, uint64_t number);
 
 #endif
